added (int) to strlen

+++ ls/hls.c
@@ -33,7 +33,7 @@ void print_directory_contents(char *dir_name, int show_hidden)
 		}
 
 		/* Convert filename to lowercase */
-		for (int i = 0; i < strlen(entry->d_name); i++)
+		for (int i = 0; i < (int)strlen(entry->d_name); i++)
 		{
 			buffer[i] = tolower(entry->d_name[i]);
 		}
